+++ storefront/all.php
@@ -10,11 +10,11 @@ $sql = 'SELECT a.*, GROUP_CONCAT(sc.name ORDER BY sc.ord SEPARATOR ", ") as cats
 $apps=$db->query($sql)->fetchAll(PDO::FETCH_ASSOC);
 
 $theme = cms_get_setting('theme','2005_v2');
-$tpl_body = __DIR__.'/templates/2005_all.twig';
+$tpl_body = dirname(__DIR__).'/themes/2005_v1/storefront/layout/2005_all.twig';
 $links = cms_load_store_links(__FILE__);
 $params = ['apps'=>$apps,'links'=>$links, 'theme_subdir' => 'storefront'];
 ob_start();
-cms_render_template($tpl_body, $params);
+cms_render_template_theme($tpl_body, '2005_v1', $params);
 $body = ob_get_clean();
 
 $tpl = cms_theme_layout('default.twig', $theme);